+++ src/main/java/com/example/rozetka_app/repositories/VideoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VideoRepository extends JpaRepository<Video, Integer> {
+public interface VideoRepository extends JpaRepository<Video, Long> {
     public List<Video> findAllByTitleLikeOrAuthorLike(String title, String author);
 }
